// frontend/src/api/generated/CalendarBookingDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type CalendarBookingDto = { booking_id: string, trip_id: string, status: string, quantity: bigint, start_date: string, end_date: string, project_id: string, project_reference: string | null, project_title: string, };